+++ packages/gatsby/src/schema/schema.js
@@ -494,8 +494,13 @@ const addCustomResolveFunctions = async ({ schemaComposer, parentSpan }) => {
             const originalFieldConfig = tc.getFieldConfig(fieldName)
             const originalTypeName = originalFieldConfig.type.toString()
             const originalResolver = originalFieldConfig.resolve
-            const fieldTypeName =
-              fieldConfig.type && fieldConfig.type.toString()
+            let fieldTypeName
+            if (fieldConfig.type) {
+              fieldTypeName = Array.isArray(fieldConfig.type)
+                ? stringifyArray(fieldConfig.type)
+                : fieldConfig.type.toString()
+            }
+
             if (
               !fieldTypeName ||
               fieldTypeName.replace(/!/g, ``) ===
@@ -742,6 +747,11 @@ const reportParsingError = error => {
     throw error
   }
 }
+
+const stringifyArray = arr =>
+  `[${arr.map(item =>
+    Array.isArray(item) ? stringifyArray(item) : item.toString()
+  )}]`
 
 // TODO: Import this directly from graphql-compose once we update to v7
 const isNamedTypeComposer = type =>